requirements.txt
pyTelegramBotAPI>=4.14.0
yt-dlp>=2023.0.0
requests>=2.31.0
python-dotenv>=1.0.0
